PrivateRoute and logout and login added
