+++ package.json
@@ -10,6 +10,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.45.0",
+    "react-icons": "^4.10.1",
     "react-router-dom": "^6.13.0",
     "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
