+++ src/layout/Main.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Outlet } from "react-router-dom";
 import Sidebar from "../Components/Sidebar/Sidebar";
-import Navbar from "../utls/Navbar/Navbar";
+import Navbar from "../utils/Navbar/Navbar";
 
 const Main = () => {
   return (
@@ -11,7 +11,7 @@ const Main = () => {
           <Sidebar />
         </div>
         <div className="w-full lg:w-[70%]  bg-white">
-          <Navbar />
+          <Navbar/>
           <Outlet />
         </div>
       </div>
