+++ src/router/PrivateRoute/PrivateRoute.jsx
@@ -1,0 +1,18 @@
+import React, { useContext } from 'react';
+import { AuthContext } from '../../Context/AuthProvider';
+import { Navigate, useLocation } from 'react-router-dom';
+
+const PrivateRoute = ({children}) => {
+    const {user, loading} = useContext(AuthContext);
+    const location = useLocation();
+    console.log(loading);
+    if (loading) {
+        return <div className='flex justify-center'>Loading ..</div>
+    }
+    if (user) {
+        return children;
+    }
+    return <Navigate to='/login' state={{from: location}} replace></Navigate>;
+};
+
+export default PrivateRoute;
